+++ getid3/module.tag.lyrics3.php
@@ -1,4 +1,5 @@
 <?php
+
 /////////////////////////////////////////////////////////////////
 /// getID3() by James Heinrich <info@getid3.org>               //
 //  available at http://getid3.sourceforge.net                 //
@@ -17,7 +18,9 @@
 
 class getid3_lyrics3 extends getid3_handler
 {
-
+	/**
+	 * @return bool
+	 */
 	public function Analyze() {
 		$info = &$this->getid3->info;
 
@@ -61,7 +64,7 @@ class getid3_lyrics3 extends getid3_handler
 
 			// Lyrics3v2, no ID3v1, no APE
 
-			$lyrics3size    = strrev(substr(strrev($lyrics3_id3v1), 9, 6)) + 6 + strlen('LYRICS200'); // LSZ = lyrics + 'LYRICSBEGIN'; add 6-byte size field; add 'LYRICS200'
+			$lyrics3size    = (int) strrev(substr(strrev($lyrics3_id3v1), 9, 6)) + 6 + strlen('LYRICS200'); // LSZ = lyrics + 'LYRICSBEGIN'; add 6-byte size field; add 'LYRICS200'
 			$lyrics3offset  = $info['filesize'] - $lyrics3size;
 			$lyrics3version = 2;
 
@@ -126,6 +129,13 @@ class getid3_lyrics3 extends getid3_handler
 		return true;
 	}
 
+	/**
+	 * @param int $endoffset
+	 * @param int $version
+	 * @param int $length
+	 *
+	 * @return bool
+	 */
 	public function getLyrics3Data($endoffset, $version, $length) {
 		// http://www.volweb.cz/str/tags.htm
 
@@ -250,6 +260,11 @@ class getid3_lyrics3 extends getid3_handler
 		return true;
 	}
 
+	/**
+	 * @param string $rawtimestamp
+	 *
+	 * @return int|false
+	 */
 	public function Lyrics3Timestamp2Seconds($rawtimestamp) {
 		if (preg_match('#^\\[([0-9]{2}):([0-9]{2})\\]$#', $rawtimestamp, $regs)) {
 			return (int) (($regs[1] * 60) + $regs[2]);
@@ -257,6 +272,11 @@ class getid3_lyrics3 extends getid3_handler
 		return false;
 	}
 
+	/**
+	 * @param array $Lyrics3data
+	 *
+	 * @return bool
+	 */
 	public function Lyrics3LyricsTimestampParse(&$Lyrics3data) {
 		$lyricsarray = explode("\r\n", $Lyrics3data['raw']['LYR']);
 		$notimestamplyricsarray = array();
@@ -288,6 +308,11 @@ class getid3_lyrics3 extends getid3_handler
 		return true;
 	}
 
+	/**
+	 * @param string $char
+	 *
+	 * @return bool|null
+	 */
 	public function IntString2Bool($char) {
 		if ($char == '1') {
 			return true;
